asdfasdfsdvsd

+++ include/function.h
@@ -1,6 +1,6 @@
 
 
-void SetVelocity(int x);
+void SetVelocity(double x);
 void drivefunctions(std::string dir, double x);
 void stap (std::string option);
 void FrontLift (std::string dir, double x, bool wait);
